# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "2.0.0",
    "title": "Volcanoes around the world v2",
    "description": "This API has been created to support assignment three in CAB230, Web Computing. It exposes a number of REST endpoints which implement CRUD operations on a database containing publicly available data about volcanoes collated by the Smithsonian Institution's Global Volcanism Program. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    },
    "schemas": {
      "MissingCountryParameterVolcanoes": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Country is a required query parameter."
          }
        }
      },
      "InvalidParametersVolcanoes": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Invalid query parameters. Only country and populatedWithin are permitted."
          }
        }
      },
...
      "VolcanoIdNotFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Volcano with ID: 99999 not found."
          }
        }
      },
      "InvalidPopulatedWithinParameterVolcanoes": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Invalid value for populatedWithin. Only: 5km,10km,30km,100km are permitted."
          }
        }
      },
      "InvalidParametersVolcano": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Invalid query parameters. Query parameters are not permitted."
          }
        }
      },
...
      "TokenExpired": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "JWT token has expired"
          }
        }
      },
      "InvalidJWT": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Invalid JWT token"
          }
        }
      },
      "MalformedAuthHeader": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Authorization header is malformed"
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "example": "This is a beautiful volcano!"
          },
          "rating": {
            "type": "integer",
            "format": "int32",
            "example": 5
          }
        }
      },
      "Photo": {
        "type": "object",
        "properties": {
          "photo_url": {
            "type": "string",
            "example": "http://example.com/photo.jpg"
          }
        }
      },
      "NoCommentsFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "No comments found for this volcano."
          }
        }
      },
      "InvalidRatingError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Invalid rating value"
          }
        }
      },
      "NoPhotosFound": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "No photos found for this volcano."
          }
        }
      },
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Unauthorized"
          }
        }
      },
...
        "InternalServerError": {
          "type": "object",
          "properties": {
            "error": {
              "type": "boolean",
              "example": true
            },
            "message": {
              "type": "string",
              "example": "Internal server error"
            }
          }
        }
      }
  },
  "paths": {
    "/countries": {
      "get": {
        "description": "Returns a list of all countries that are associated with one or more volcanoes, ordered alphabetically.",
        "responses": {
          "200": {
            "description": "An array of countries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": ["Algeria", "Antarctica", "Argentina", "Armenia", "Australia"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidParametersVolcanoes"
                }
              }
            }
          }
        }
      }
    },
    "/volcanoes": {
      "get": {
        "description": "Fetches volcanoes based on country and optionally filtered by population within certain distances.",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": true,
            "description": "The country to filter volcanoes.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populatedWithin",
            "in": "query",
            "description": "Filter volcanoes based on population within a specified radius.",
            "schema": {
              "type": "string",
              "enum": ["5km", "10km", "30km", "100km"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of volcanoes.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": "1"
                      },
                      "name": {
                        "type": "string",
                        "example": "Abu"
                      },
                      "country": {
                        "type": "string",
                        "example": "Japan"
                      },
                      "region": {
                        "type": "string",
                        "example": "Japan, Taiwan, Marianas"
                      },
                      "subregion": {
                        "type": "string",
                        "example": "Honshu"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing country query parameter or query parameters invalid. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MissingCountryParameterVolcanoes"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidParametersVolcanoes"
                    },
                    {
                      "$ref": "#/components/schemas/InvalidPopulatedWithinParameterVolcanoes"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/volcano/{id}": {
      "get": {
        "description": "Returns an object containing data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for 5km, 10km, 30km and 100km is also provided. To test this using Swagger, click the green 'Authorize' button at the top of this page to enter your JWT token. A JWT token can be obtained by logging in. The path parameter (id) is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the volcano",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object containing id, name, country, region, subregion, last eruption, summit, elevation, latitude and longitude data for the queried volcano. If a valid JWT token is sent in the header of the request, population data for 5km, 10km, 30km and 100km is also provided. An example of one object (with an authenticated request) is shown below.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Abu"
                    },
                    "country": {
                      "type": "string",
                      "example": "Japan"
                    },
                    "region": {
                      "type": "string",
                      "example": "Japan, Taiwan, Marianas"
                    },
                    "subregion": {
                      "type": "string",
                      "example": "Honshu"
                    },
                    "last_eruption": {
                      "type": "string",
                      "example": "6850 BCE"
                    },
                    "summit": {
                      "type": "number",
                      "example": 641
                    },
                    "elevation": {
                      "type": "number",
                      "example": 2103
                    },
                    "latitude": {
                      "type": "string",
                      "example": "34.5000"
                    },
                    "longitude": {
                      "type": "string",
                      "example": "131.6000"
                    },
                    "population_5km": {
                      "type": "number",
                      "example": 3597
                    },
                    "population_10km": {
                      "type": "number",
                      "example": 9594
                    },
                    "population_30km": {
                      "type": "number",
                      "example": 117805
                    },
                    "population_100km": {
                      "type": "number",
                      "example": 4071152
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidParametersVolcano"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InvalidJWT"
                    },
                    {
                      "$ref": "#/components/schemas/TokenExpired"
                    },
                    {
                      "$ref": "#/components/schemas/MalformedAuthHeader"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Volcano ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/VolcanoIdNotFound"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/volcanoes/{id}/comments": {
      "get": {
        "tags": ["Volcano Comments"],
        "summary": "Retrieve comments and average rating for a specific volcano",
        "description": "Fetches all comments and calculates the average rating for the given volcano ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the volcano"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of comments and average rating",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Comment"
                      }
                    },
                    "averageRating": {
                      "type": "number",
                      "format": "float",
                      "example": 4.5
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No comments found for this volcano.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoCommentsFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Volcano Comments"],
        "summary": "Allows users to comment on and rate volcanoes",
        "description": "Adds a user's comment and rating to a volcano.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the volcano"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Comment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment added successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid rating value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidRatingError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": " #/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/volcanoes/{id}/photos": {
      "get": {
        "tags": ["Volcano Photos"],
        "summary": "Retrieve photos for a specific volcano",
        "description": "Fetches all photos related to the given volcano ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the volcano"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of photo URLs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Photo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No photos found for this volcano.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoPhotosFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Volcano Photos"],
        "summary": "Allows users to upload photos of volcanoes",
        "description": "Adds a photo URL to the database under a specific volcano ID, requires user authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the volcano"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "123"
                  },
                  "url": {
                    "type": "string",
                    "example": "http://example.com/photo.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Photo added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Photo added successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, no or invalid authentication details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "Volcano not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolcanoIdNotFound"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    }
  }
}